+++ NEWS.md
@@ -8,8 +8,7 @@ Cite all releases with doi [10.5281/zenodo.8226419](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
-* Add new function `load_schema_table()`.
-* Add new function `schema_dm()`.
+* Add new functions `load_schema_table()`, `schema_dm()`, and `pl_destroy()`.
 
 
 # PFUPipelineTools 0.1.8 (2023-12-21) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10420449.svg)](https://doi.org/10.5281/zenodo.10420449)